renamed setup.py to session.py
