+++ tests/parity/conftest.py
@@ -1,6 +1,6 @@
 import pytest
 
-from ether_sql.setup import Session
+from ether_sql.session import Session
 from ether_sql.models import base
 
 pytest_plugins = [
